api adjusted to three levels of access: visitor, user, admin
