+++ src/api/admin_article_api.rs
@@ -2,12 +2,11 @@ use sapper::{ SapperModule, SapperRouter, Response, Request, Result as SapperRes
 use sapper_std::{ PathParams, QueryParams, JsonParams };
 use serde_json;
 
-use super::super::{ EditArticle, ModifyPublish, Articles,
-                    NewArticle, ArticleList, Postgresql };
+use super::super::{ NewArticle, Articles, Postgresql, EditArticle, ArticleList, ModifyPublish };
 
-pub struct Article;
+pub struct AdminArticle;
 
-impl Article {
+impl AdminArticle {
     fn create_article(req: &mut Request) -> SapperResult<Response> {
         let body: NewArticle = get_json_params!(req);
         let pg_pool = req.ext().get::<Postgresql>().unwrap().get().unwrap();
@@ -63,56 +62,12 @@ impl Article {
         res_json!(res)
     }
 
-    fn view_article(req: &mut Request) -> SapperResult<Response> {
-        let params = get_query_params!(req);
-        let article_id = t_param_parse!(params, "id", i32);
-        let pg_pool = req.ext().get::<Postgresql>().unwrap().get().unwrap();
-
-        let res = match Articles::query_article(&pg_pool, article_id, false) {
-            Ok(data) => {
-                json!({
-                    "status": true,
-                    "data": data
-                })
-            }
-            Err(err) => {
-                json!({
-                    "status": false,
-                    "error": err
-                })
-            }
-        };
-        res_json!(res)
-    }
-
     fn admin_list_all_article(req: &mut Request) -> SapperResult<Response> {
         let params = get_query_params!(req);
         let limit = t_param_parse!(params, "limit", i64);
         let offset = t_param_parse!(params, "offset", i64);
         let pg_pool = req.ext().get::<Postgresql>().unwrap().get().unwrap();
         let res = match ArticleList::query_list_article(&pg_pool, limit, offset, true) {
-            Ok(data) => {
-                json!({
-                    "status": true,
-                    "data": data
-                })
-            }
-            Err(err) => {
-                json!({
-                    "status": false,
-                    "error": err
-                })
-            }
-        };
-        res_json!(res)
-    }
-
-    fn list_all_article(req: &mut Request) -> SapperResult<Response> {
-        let params = get_query_params!(req);
-        let limit = t_param_parse!(params, "limit", i64);
-        let offset = t_param_parse!(params, "offset", i64);
-        let pg_pool = req.ext().get::<Postgresql>().unwrap().get().unwrap();
-        let res = match ArticleList::query_list_article(&pg_pool, limit, offset, false) {
             Ok(data) => {
                 json!({
                     "status": true,
@@ -137,7 +92,7 @@ impl Article {
 
         let res = match Articles::edit_article(&pg_pool, body) {
             Ok(num_update) => {
-                 json!({
+                json!({
                     "status": true,
                     "num_update": num_update
                 })
@@ -176,7 +131,7 @@ impl Article {
     }
 }
 
-impl SapperModule for Article {
+impl SapperModule for AdminArticle {
     fn before(&self, _req: &mut Request) -> SapperResult<()> {
         Ok(())
     }
@@ -187,28 +142,23 @@ impl SapperModule for Article {
 
     fn router(&self, router: &mut SapperRouter) -> SapperResult<()> {
         // http get /article/admin/view id==4
-        router.get("/article/admin/view", Article::admin_view_article);
+        router.get("/article/admin/view", AdminArticle::admin_view_article);
 
         // http get /article/admin/view_all limit==5
-        router.get("/article/admin/view_all", Article::admin_list_all_article);
-
-        // http get /article/view id==4
-        router.get("/article/view", Article::view_article);
-
-        // http get /article/view_all limit==5
-        router.get("/article/view_all", Article::list_all_article);
+        router.get("/article/admin/view_all", AdminArticle::admin_list_all_article);
 
         // http post /article/new title=something raw_content=something
-        router.post("/article/new", Article::create_article);
-
-        // http post /article/publish id:=5 published:=true
-        router.post("/article/publish", Article::update_publish);
+        router.post("/article/new", AdminArticle::create_article);
 
         // http post /article/delete/3
-        router.post("/article/delete/:id", Article::delete_article);
+        router.post("/article/delete/:id", AdminArticle::delete_article);
 
         // http post /article/edit id:=1 title=something raw_content=something
-        router.post("/article/edit", Article::edit_article);
+        router.post("/article/edit", AdminArticle::edit_article);
+
+        // http post /article/publish id:=5 published:=true
+        router.post("/article/publish", AdminArticle::update_publish);
+
         Ok(())
     }
 }
