+++ src/api/mod.rs
@@ -1,7 +1,15 @@
-pub mod blog_user_api;
-pub mod blog_article_api;
-pub mod blog_tag_api;
+pub mod visitor_api;
+pub mod user_api;
 
-pub use self::blog_article_api::Article;
-pub use self::blog_user_api::User;
-pub use self::blog_tag_api::Tag;
+pub use self::visitor_api::Visitor;
+pub use self::user_api::User;
+
+pub mod admin_user_api;
+pub mod admin_tag_api;
+pub mod admin_article_api;
+
+pub use self::admin_user_api::AdminUser;
+pub use self::admin_tag_api::Tag;
+pub use self::admin_article_api::AdminArticle;
+
+
