+++ src/lib.rs
@@ -38,7 +38,9 @@ pub(crate) use models::{ NewTag, Tags, TagCount };
 pub(crate) use util::{ sha3_256_encode, random_string, markdown_render, get_password };
 pub use util::{ create_redis_pool, RedisPool, Redis };
 pub use util::{ create_pg_pool, Postgresql };
-pub use api::Article;
+pub use api::Visitor;
 pub use api::User;
+pub use api::AdminArticle;
 pub use api::Tag;
+pub use api::AdminUser;
 pub use web::ArticleWeb;
